CSI Userland - cp EXAMPLE vagrant files if vagrant.yaml does not exist in respective userland folder per vagrant_rsync_userland_configs.lst in order to get through an installation - slight_tewak to omit files other than vagrant.yaml

+++ install.sh
@@ -23,7 +23,7 @@ fi
 vagrant plugin install vagrant-reload
 
 cd /csi && cat ./vagrant_rsync_userland_configs.lst | while read userland_config; do
-  if [[ ! -e $userland_config ]]; then
+  if [[ `basename ${userland_config}` == 'vagrant.yaml' && ! -e $userland_config ]]; then
     cp $userland_config.EXAMPLE $userland_config
   fi
 done
